Updated SmiteDialog based on AwardAndSmiteForm addition
