+++ src/modules/dashboard/components/AwardAndSmiteDialogForm/AwardAndSmiteDialogForm.tsx
@@ -97,7 +97,7 @@ const SmiteDialogForm = ({
     hasRegisteredProfile && userHasRole(domainRoles, ColonyRole.Arbitration);
 
   const [userHasPermission, onlyForceAction] = useDialogActionPermissions(
-    colony.colonyAddress,
+    colonyAddress,
     hasRoles,
     isVotingExtensionEnabled,
     values.forceAction,
@@ -173,7 +173,7 @@ const SmiteDialogForm = ({
   >(
     (option) => {
       const value = option ? option.value : undefined;
-      const domain = colony.domains.find(
+      const domain = domains.find(
         ({ ethDomainId }) => Number(value) === ethDomainId,
       ) as OneDomain;
       return (
@@ -186,7 +186,7 @@ const SmiteDialogForm = ({
         </div>
       );
     },
-    [values, colonyAddress, colony.domains],
+    [values, colonyAddress, domains],
   );
 
   useEffect(() => {
