--- src/modules/dashboard/components/SmiteDialog/SmiteDialogForm.css
@@ -1,86 +1,0 @@
-
-@value wideButton: 160px;
-
-.inputContainer {
-  margin-bottom: 20px;
-  width: 320px;
-  position: relative;
-}
-
-.inputContainer label,
-.domainSelects label {
-  font-size: var(--size-smallish);
-  color: var(--dark);
-}
-
-.inputContainer label {
-  margin-top: -20px;
-  height: auto;
-  width: auto;
-  clip: auto;
-}
-
-.inputContainer input {
-  padding-right: 30px;
-  font-weight: var(--weight-normal);
-  color: var(--dark);
-}
-
-.inputContainer button {
-  left: -6px;
-}
-
-.singleUserContainer label {
-  margin-bottom: 15px;
-  color: var(--dark);
-}
-
-.singleUserContainer > div {
-  padding-left: 0;
-}
-
-.noPermissionFromMessage {
-  margin-bottom: 20px;
-  font-size: var(--size-tiny);
-  font-weight: var(--weight-bold);
-  line-height: 18px;
-  color: var(--danger);
-}
-
-.noPermissionFromMessage > div {
-  vertical-align: bottom;
-}
-
-.modalHeading {
-  padding: 32px 0 20px;
-}
-
-.headingContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.inputText {
-  margin-top: 5px;
-  float: right;
-  font-size: var(--size-tiny);
-  font-weight: var(--weight-bold);
-  color: var(--temp-grey-blue-7);
-}
-
-.activeItem div {
-  padding: 0;
-}
-
-.motionVoteDomain {
-  margin: 6px 0;
-}
-
-.percentageSign {
-  position: absolute;
-  top: 6px;
-  right: 0;
-  font-size: var(--size-medium);
-  color: var(--dark);
-}
