--- src/modules/dashboard/components/SmiteDialog/SmiteDialogForm.css.d.ts
@@ -1,11 +1,0 @@
-export const wideButton: string;
-export const inputContainer: string;
-export const domainSelects: string;
-export const singleUserContainer: string;
-export const noPermissionFromMessage: string;
-export const modalHeading: string;
-export const headingContainer: string;
-export const inputText: string;
-export const activeItem: string;
-export const motionVoteDomain: string;
-export const percentageSign: string;
