--- src/modules/dashboard/components/SmiteDialog/TeamDropdownItem.css.d.ts
@@ -1,3 +1,0 @@
-export const main: string;
-export const headingWrapper: string;
-export const color: string;
